consistent-ize ogUrl, move to utils
